Cloud Functions/FetchTaxiLocation/requirements.txt
# Function dependencies, for example:
# package>=version
firebase_admin
utm
datetime
google-cloud-storage
pandas
requests